+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to the MoveApps R SDK will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- "Restore default settings" (shiny) button next to the "Store settings" button, which reloads the app with its default settings (new exported function `restoreDefaultSettings()`)
+
 ## [v1.0.3] - 2025-12-16
 
 ### Changed